+++ roop/uis/__layouts__/default.py
@@ -1,6 +1,6 @@
 import gradio
 
-from roop.uis.__components__ import settings, source, target, preview, trim_frame, face_selector, output
+from roop.uis.__components__ import settings, source, target, preview, trim_frame, face_analyser, face_selector, output
 
 
 def render() -> gradio.Blocks:
@@ -15,6 +15,7 @@ def render() -> gradio.Blocks:
                 preview.render()
                 trim_frame.render()
                 face_selector.render()
+                face_analyser.render()
         with gradio.Row():
             output.render()
     return layout
@@ -27,4 +28,5 @@ def listen() -> None:
     preview.listen()
     trim_frame.listen()
     face_selector.listen()
+    face_analyser.listen()
     output.listen()
